loadbalancer: Add cell package to tie everything together
Add a cell package that contains the root module for load-balancing
that ties everything together.
