+++ pkg/loadbalancer/README.md
@@ -55,16 +55,10 @@ The core source files are:
 - config.go: Configuration structs. `Config` is the main configuration, `ExtConfig` bridges from
   `option.DaemonConfig` to avoid direct references to it (to be removed eventually) and `TestConfig`
   allows tests to tweak the behavior.
-- writer.go: API for modifying the tables while maintaining cross-table references
-- bpf_reconciler.go: Reconciliation from frontends to BPF maps
-- reflector.go: Reflects Kubernetes Service, Pod and EndpointSlices to the tables (via Writer)
-- cmds.go: Implementation of lb/* script commands for inspection and testing
-- lbmaps.go: API towards load-balancing BPF maps
-
-Additionally:
-- adapters.go: Adaption to the older load-balancing APIs (to be removed)
-- node_addr_reconciler.go: Refreshes NodePort frontends on changes to node addresses
-- healthserver.go: Implements the NodePort health check support
+- writer: API for modifying the tables while maintaining cross-table references
+- reconciler: Reconciliation from frontends to BPF maps
+- maps: "LBMaps" wrapper around BPF maps
+- reflectors: Reflects Kubernetes Service, Pod and EndpointSlices to the tables (via Writer)
 
 ## Inspecting state
 
@@ -85,7 +79,7 @@ The new architecture makes it easier to write integration tests due to the decou
 integration test for a data source can depend just on the tables & writer and verify that the
 tables are correctly populated without having to mock other features or the BPF operations.
 
-An "end-to-end" style test can be found in `script_test.go` that tests going from Kubernetes
+An "end-to-end" style test can be found in `tests/script_test.go` that tests going from Kubernetes
 objects specified in YAML to the BPF map contents using script tests in `testdata`. This is
 the main way of testing the control-plane.
 
@@ -94,7 +88,7 @@ and `pkg/ciliumenvoyconfig/script_test.go`. For more info on script tests see
 https://docs.cilium.io/en/latest/contributing/development/hive/#testing-with-hive-script and
 https://docs.cilium.io/en/latest/contributing/development/statedb/#script-commands.
 
-For quick feedback loop you can use the `watch.sh` script to run a script test when the
+For quick feedback loop you can use the `watch.sh` script (in tests/) to run a script test when the
 txtar file changes.
 
 IMPORTANT: When adding new test cases it's important to stress test to make sure the new
@@ -107,8 +101,9 @@ Tests can be executed normally with "go test":
 
 To run the privileged tests that use the real BPF maps:
 ```
+  $ cd tests
   $ go test -c
-  $ PRIVILEGED_TESTS=1 sudo -E ./experimental.test -test.run . -test.v -test.count 1
+  $ PRIVILEGED_TESTS=1 sudo -E ./tests.test -test.run . -test.v -test.count 1
 ```
 
 ## Benchmarking
