+++ pkg/loadbalancer/cell/adapters.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package experimental
+package cell
 
 import (
 	"context"
