+++ pkg/loadbalancer/maps/cell.go
@@ -4,7 +4,13 @@
 package maps
 
 import (
+	"errors"
+	"sync"
+
 	"github.com/cilium/hive/cell"
+	"golang.org/x/sys/unix"
+
+	"github.com/cilium/cilium/pkg/netns"
 )
 
 // Provides [LBMap] a wrapper around the load-balancing BPF maps
@@ -17,4 +23,16 @@ var Cell = cell.Module(
 
 	// Provide the 'lb/' script commands for debugging and testing.
 	cell.Provide(scriptCommands),
+
+	// Provide [HaveNetNSCookieSupport] to probe for netns cookie support.
+	cell.Provide(NetnsCookieSupportFunc),
 )
+
+type HaveNetNSCookieSupport func() bool
+
+func NetnsCookieSupportFunc() HaveNetNSCookieSupport {
+	return sync.OnceValue(func() bool {
+		_, err := netns.GetNetNSCookie()
+		return !errors.Is(err, unix.ENOPROTOOPT)
+	})
+}
